Refactor frontend structure and integrate API
